Moved a function into another file in the same pkg to simplify the project.
